+++ src/main/java/org/openlmis/referencedata/errorhandling/RefDataErrorHandling.java
@@ -62,6 +62,8 @@ public class RefDataErrorHandling extends BaseHandler {
     CONSTRAINT_MAP.put("unq_programid_orderableid_orderableversionid",
         OrderableMessageKeys.ERROR_PROGRAMS_DUPLICATED);
     CONSTRAINT_MAP.put("unq_supply_partner_code", SupplyPartnerMessageKeys.ERROR_CODE_DUPLICATED);
+    CONSTRAINT_MAP.put("unq_supply_partner_association_programId_supervisoryNodeId",
+        SupplyPartnerMessageKeys.ERROR_ASSOCIATION_DUPLICATED);
   }
 
   /**