+++ src/main/resources/db/migration/20181108091545789__add_unique_index_on_programId_supervisoryNodeId_in_supply_partner_associations.sql
@@ -1,0 +1,2 @@
+CREATE UNIQUE INDEX unq_supply_partner_association_programId_supervisoryNodeId
+ON supply_partner_associations (programId, supervisoryNodeId, supplyPartnerId);