+++ CHANGELOG.md
@@ -29,6 +29,8 @@ New functionality added in a backwards-compatible manner:
   * Added ability to find FTAPs for several facility types
   * Added ability to retrieve all supervising facilities for the given supervisory node.
 * [OLMIS-5143](https://openlmis.atlassian.net/browse/OLMIS-5143): Added `extraData` field to supervisory node resource
+* [OLMIS-5614](https://openlmis.atlassian.net/browse/OLMIS-5614): Provided validation for supply partner resource
+  * Added bidirectional relation between regular and partner nodes
   
 Improvements:
 * [OLMIS-4295](https://openlmis.atlassian.net/browse/OLMIS-4295): Updated checkstyle to use newest google style.